image_project/requirements.txt
fastapi
pydantic
numpy
opencv-python-headless
uvicorn
Pillow
pytest